src/requirements.txt
amqp==5.0.9
asgiref==3.5.0
aspose-words==21.11.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dj-static==0.0.6
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.0.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-fernet-fields==0.6
django-filter==21.1
django-grappelli==3.0.2
django-health-check==3.16.5
djangorestframework==3.13.1
drf-yasg2==1.19.4
markdown
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==19.9.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonformatter
jsonify
kombu==5.2.3
MarkupSafe==2.0.1
minio==7.1.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pdfkit==1.0.0
protobuf==3.19.4
psutil
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.0.7
python-dotenv==0.19.2
python-interface==1.6.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
uritemplate==4.1.1
urllib3==1.26.8
uwsgi
vine==5.0.0
Werkzeug
wget
xmltodict==0.12.0
zipp==3.7.0